(maint) Embed the MSI in the nuget pkgs
Previously the PDK, Bolt and Puppet-Agent packages relied on downloading the MSI
files from well known URLs, however the URLs will be moving soon which caused
all of the packages to be rebuilt.  To make the packages resistant to URL
changes in the future these packages will now contain the actual MSIs needed to
install, instead of the URL to the MSIs.