(maint) PDK Chocolatey package should not vendor the MSI
Previously in commit c96fe7f4 the PDK packaging was modified to vendor the MSI,
however the PDK MSI is growing quite large and will soon exceed to the 200MB
limit.  This means it cannot be vendored into the package.  This commit
partially reverts the changes in commit c96fe7f4 and instead uses the new URLs
for the PDK location instead of the old broken URI.