(maint) PDK Puppet Agent package should not vendor the MSI
Previously in commit c96fe7f4 the Puppet Agent packaging was modified to vendor
the MSI, however the combined 32 and 64 but package size is growing quite large
and may exceed to the 200MB limit.  This means they cannot be vendored into the
package.  This commit partially reverts the changes in commit c96fe7f4 and
instead uses the new URLs for the Puppet Agent location instead of the old
broken URI.